Refactoring.
Renamed Submission.finalizationToken to stateChanged because I plan to use it to detect all state changes and not just completion/cancellation.
I also changed it to reset itself as a new promise after the previous was fulfilled. This might not be such a good idea, though. I want to use it in the queue manager to handle the start/pause/stop functionality within a given timeout. We'll see...